requirements.txt
boto3==1.9.86
botocore==1.12.86
certifi==2019.3.9
chardet==3.0.4
Django==1.11
django-filter==2.1.0
django-pgcrypto==1.4.0
django-pgcrypto-fields==2.5.1
django-rest-framework==0.1.0
djangorestframework==3.9.1
docutils==0.14
futures==2.2.0
geoip2==2.9.0
idna==2.8
jmespath==0.9.3
kafka-python==1.4.6
maxminddb==1.4.1
Pillow==2.2.1
psycopg2-binary==2.7.7
pycrypto==2.6.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
urllib3==1.24.1
